update: update now only user who create can modify

+++ ProjectManagment/Application/ProjectsUsers/Commands/DeleteUserFromProjectCommand.cs
@@ -11,11 +11,14 @@ namespace Application.ProjectsUsers.Commands;
 public class DeleteUserFromProjectCommand : IRequest<Result<ProjectUser, ProjectUserException>>
 {
     public required Guid ProjectId { get; init; }
+    public required Guid UserIdWhoCreated { get; init; }
     public required Guid UserId { get; init; }
 }
 
 public class
-    DeleteUserFromProjectCommandHandler(IProjectUserRepository repository)
+    DeleteUserFromProjectCommandHandler(IProjectUserRepository repository,
+        IProjectRepository projectRepository,
+        IUserRepository userRepository)
     : IRequestHandler<DeleteUserFromProjectCommand,
         Result<ProjectUser, ProjectUserException>>
 {
@@ -24,12 +27,31 @@ public class
     {
         var projectId = new ProjectId(request.ProjectId);
         var userId = new UserId(request.UserId);
+        var userIdWhoCreated = new UserId(request.UserIdWhoCreated);
         var exsistingProjectUser = await repository.GetByIds(projectId, userId, cancellationToken);
+        var existingProject = await projectRepository.GetById(projectId, cancellationToken);
+        var existingUser = await userRepository.GetById(userIdWhoCreated, cancellationToken);
 
-        return await exsistingProjectUser.Match(
-            async pu => await DeleteUser(pu, cancellationToken),
-            () => Task.FromResult<Result<ProjectUser, ProjectUserException>>(
-                new ProjectUserNotFound(projectId, userId)));
+        return await existingProject.Match(
+            async p =>
+            {
+                return await existingUser.Match(
+                    async u =>
+                    {
+                        if (p.UserId != userIdWhoCreated || u.Role.Name != "Admin")
+                        {
+                            return await Task.FromResult<Result<ProjectUser, ProjectUserException>>(
+                                new UserNotEnoughPremission(projectId, userId));
+                        }
+                        return await exsistingProjectUser.Match(
+                            async pu => await DeleteUser(pu, cancellationToken),
+                            () => Task.FromResult<Result<ProjectUser, ProjectUserException>>(
+                                new ProjectUserNotFound(projectId, userId)));
+                    },
+                    () => Task.FromResult<Result<ProjectUser, ProjectUserException>>(
+                        new UserNotFound(projectId, userId)));
+            },
+            () => Task.FromResult<Result<ProjectUser, ProjectUserException>>(new ProjectNotFound(projectId, userId)));
     }
 
     private async Task<Result<ProjectUser, ProjectUserException>> DeleteUser(ProjectUser projectUser,
